fix(driver): avoid `ppm_consumer_t` being defined twice.
It is `typedef`ed both in ppm.h and in ppm_events.h, both included in kmod main.c.
