+++ driver/CMakeLists.txt
@@ -115,6 +115,7 @@ set(DRIVER_SOURCES
 	systype_compat.h
 	ppm_tp.h
 	ppm_tp.c
+	ppm_consumer.h
 )
 
 foreach(FILENAME IN LISTS DRIVER_SOURCES)
